ioteca_api/requirements.txt
# cx-Oracle==5.1.3
Django==1.9.6
django-cors-headers==1.1.0
djangorestframework==3.3.3
fpdf==1.7.2
Pillow==3.1.1
